requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-filter
psycopg2-binary